% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_datasets.R
\name{compile_ctd}
\alias{compile_ctd}
\title{Binding individual ctd datasets}
\usage{
compile_ctd(
  acc_path = "./access_files",
  dir = "./compiled_raw_datasets",
  comp_file = "ctd_all_raw_pre-qc",
  parallel = FALSE
)
}
\arguments{
\item{acc_path}{directory where the SMRU Microsoft Access files can be found for querying. Default is "./access_files"}

\item{dir}{directory where combined data.frames can be written. Default is "./compiled_raw_datasets"}

\item{comp_file}{prefix for writing the combined ctd data.frame to a CSV. Default is "ctd_all_raw_pre-qc"}

\item{parallel}{logical - should the data be processed in parallel?}
}
\value{
CSV of the combined ctd dataset for all available campaigns
}
\description{
Binding individual ctd datasets
}
\examples{
\dontrun{
compile_ctd()
}
}
