% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dive_metrics.R
\name{compute_dive_metrics}
\alias{compute_dive_metrics}
\title{Computing elephant seal dive effort metrics}
\usage{
compute_dive_metrics(
  out_dir = "./processed_datasets/",
  in_dive = "./processed_datasets/loc_ssm_dive.csv",
  out_dive = "dive_metrics",
  dep_min = 15,
  dep_max = 1200,
  dur_min = 300,
  dur_max = 2800,
  surf_max = 300,
  append = TRUE,
  chunk_prop = 0.1
)
}
\arguments{
\item{out_dir}{destination directory for outputs.}

\item{in_dive}{file path for the location processed dive dataset.}

\item{out_dive}{prefix for the output dive metrics file name.}

\item{dep_min}{threshold for minimum dive depth (metres)}

\item{dep_max}{threshold for maximum dive depth (metres)}

\item{dur_min}{threshold for minimum dive duration (seconds)}

\item{dur_max}{threshold for maximum dive duration (seconds)}

\item{surf_max}{threshold for maximum surface interval (seconds)}

\item{append}{logical. Would you like this dive metrics dataset to be appended to a previously processed one?}

\item{chunk_prop}{relative size of data chunks for splitting the dataset during processing. Value must be (0,1]}
}
\value{
a csv file with dive effort metrics computed for each dive
}
\description{
Computing elephant seal dive effort metrics
}
\examples{
\dontrun{

compute_dive_metrics()

}

}
