% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SMRU_files.R
\name{get_SMRU_files}
\alias{get_SMRU_files}
\title{Download SRDL files from the SMRU Data Portal}
\usage{
get_SMRU_files(
  dir = NULL,
  campaigns,
  providers = NULL,
  replace = FALSE,
  compiled_fpath = NULL,
  unpack = TRUE,
  unpack_path = NULL
)
}
\arguments{
\item{dir}{directory to which data files are downloaded (default is "./downloaded_campaigns/")}

\item{campaigns}{either (1) a 2-column data.frame containing a column named "campaigns" with rows containing the SMRU campaign number (e.g. ct119) and a second column called "provider" with rows containing the provider (e.g. imos), else, (2) a character vector with the SMRU campaign number(s) to be downloaded. Note: in this case "provider cannot be left NULL}

\item{providers}{if "campaigns" input is a character vector, providers should be a character vector containing the data provider (e.g. imos). Note must be the same length as vector for "campaigns"}

\item{replace}{logical - should files with the same name be overwritten?}

\item{compiled_fpath}{path to previously compiled data. If data.sets have been previously downloaded and compiled using this package, specifying the path of the compiled csv here can be used to skip downloads of these data. Default is "./compiled_raw_datasets/loc_all_raw_pre-qc.txt"}

\item{unpack}{logical - should the compressed file downloads be unzipped?}

\item{unpack_path}{directory to which unzipped files are written. Default is directory is "./access_files/"}
}
\value{
collection of zipped or unzipped Microsoft Access files in the specified directory, and containing the requested SRDL deployment data
}
\description{
Download SRDL files from the SMRU Data Portal
}
\examples{

campaigns <- c("ct164", "ct79", "ct91")
providers <- c("imos", "imos", "imos")

get_SMRU_files(campaigns = campaigns, providers = providers)

}
