% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaTracks-package.R
\docType{package}
\name{seaTracks-package}
\alias{seaTracks}
\alias{seaTracks-package}
\title{seaTracks: Dowloading, compiling and processing satellite biotelemetry data from the Sea Mammal Research Unit (SMRU) Data Portal}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: David Green \email{davo.b.green@gmail.com} (\href{https://orcid.org/0000-0002-0346-3129}{ORCID})

}
\keyword{internal}
